requirements.txt
torch
numpy
python-chess
zstandard
streamlit
